Fixing compilation of vector_for_each test
