+++ hpx/traits.hpp
@@ -153,7 +153,13 @@ namespace hpx { namespace traits
     struct segmented_iterator_traits;
 
     template <typename Iterator, typename Enable = void>
+    struct is_segmented_iterator;
+
+    template <typename Iterator, typename Enable = void>
     struct segmented_local_iterator_traits;
+
+    template <typename T, typename Enable = void>
+    struct projected_iterator;
 
     ///////////////////////////////////////////////////////////////////////////
     template <typename T, typename Enable = void>
