+++ hpx/traits/segmented_iterator_traits.hpp
@@ -18,6 +18,11 @@ namespace hpx { namespace traits
         typedef std::false_type is_segmented_iterator;
     };
 
+    template <typename Iterator, typename Enable>
+    struct is_segmented_iterator
+      : segmented_iterator_traits<Iterator>::is_segmented_iterator
+    {};
+
     ///////////////////////////////////////////////////////////////////////////
     // traits allowing to distinguish iterators which have a purely local
     // representation
@@ -52,170 +57,11 @@ namespace hpx { namespace traits
     };
 
     ///////////////////////////////////////////////////////////////////////////
-    namespace functional
+    template <typename T, typename Enable>
+    struct projected_iterator
     {
-        struct segmented_iterator_segment
-        {
-            template <typename Iterator>
-            struct apply
-            {
-                typedef typename segmented_iterator_traits<
-                        Iterator
-                    >::segment_iterator type;
-
-                template <typename T>
-                struct result;
-
-                template <typename This, typename Iter>
-                struct result<This(Iter)>
-                {
-                    typedef typename apply::type type;
-                };
-
-                template <typename Iter>
-                typename result<segmented_iterator_segment(Iter)>::type
-                operator()(Iter iter) const
-                {
-                    return segmented_iterator_traits<Iterator>::segment(iter);
-                };
-            };
-        };
-
-        struct segmented_iterator_local
-        {
-            template <typename Iterator>
-            struct apply
-            {
-                typedef typename segmented_iterator_traits<
-                        Iterator
-                    >::local_iterator type;
-
-                template <typename T>
-                struct result;
-
-                template <typename This, typename Iter>
-                struct result<This(Iter)>
-                {
-                    typedef typename apply::type type;
-                };
-
-                template <typename Iter>
-                typename result<segmented_iterator_local(Iter)>::type
-                operator()(Iter iter) const
-                {
-                    return segmented_iterator_traits<Iterator>::local(iter);
-                };
-            };
-        };
-
-        struct segmented_iterator_begin
-        {
-            template <typename Iterator>
-            struct apply
-            {
-                typedef typename segmented_iterator_traits<
-                        Iterator
-                    >::local_iterator type;
-
-                template <typename T>
-                struct result;
-
-                template <typename This, typename SegIter>
-                struct result<This(SegIter)>
-                {
-                    typedef typename apply::type type;
-                };
-
-                template <typename SegIter>
-                typename result<segmented_iterator_begin(SegIter)>::type
-                operator()(SegIter iter) const
-                {
-                    return segmented_iterator_traits<Iterator>::begin(iter);
-                };
-            };
-        };
-
-        struct segmented_iterator_end
-        {
-            template <typename Iterator>
-            struct apply
-            {
-                typedef typename segmented_iterator_traits<
-                        Iterator
-                    >::local_iterator type;
-
-                template <typename T>
-                struct result;
-
-                template <typename This, typename SegIter>
-                struct result<This(SegIter)>
-                {
-                    typedef typename apply::type type;
-                };
-
-                template <typename SegIter>
-                typename result<segmented_iterator_end(SegIter)>::type
-                operator()(SegIter iter) const
-                {
-                    return segmented_iterator_traits<Iterator>::end(iter);
-                };
-            };
-        };
-
-        struct segmented_iterator_local_begin
-        {
-            template <typename Iterator>
-            struct apply
-            {
-                typedef typename segmented_iterator_traits<
-                        Iterator
-                    >::local_raw_iterator type;
-
-                template <typename T>
-                struct result;
-
-                template <typename This, typename LocalSegIter>
-                struct result<This(LocalSegIter)>
-                {
-                    typedef typename apply::type type;
-                };
-
-                template <typename LocalSegIter>
-                typename result<segmented_iterator_local_begin(LocalSegIter)>::type
-                operator()(LocalSegIter iter) const
-                {
-                    return segmented_iterator_traits<Iterator>::begin(iter);
-                };
-            };
-        };
-
-        struct segmented_iterator_local_end
-        {
-            template <typename Iterator>
-            struct apply
-            {
-                typedef typename segmented_iterator_traits<
-                        Iterator
-                    >::local_raw_iterator type;
-
-                template <typename T>
-                struct result;
-
-                template <typename This, typename LocalSegIter>
-                struct result<This(LocalSegIter)>
-                {
-                    typedef typename apply::type type;
-                };
-
-                template <typename LocalSegIter>
-                typename result<segmented_iterator_local_end(LocalSegIter)>::type
-                operator()(LocalSegIter iter) const
-                {
-                    return segmented_iterator_traits<Iterator>::end(iter);
-                };
-            };
-        };
-    }
+        typedef typename hpx::util::decay<T>::type type;
+    };
 }}
 
 #endif
