+++ hpx/util/zip_iterator.hpp
@@ -8,10 +8,11 @@
 #define HPX_UTIL_ZIP_ITERATOR_MAY_29_2014_0852PM
 
 #include <hpx/util/tuple.hpp>
+#include <hpx/traits/segmented_iterator_traits.hpp>
 #include <hpx/util/detail/pack.hpp>
 #include <hpx/util/result_of.hpp>
+#include <hpx/util/functional/segmented_iterator_helpers.hpp>
 #include <hpx/runtime/serialization/serialize_sequence.hpp>
-#include <hpx/traits/segmented_iterator_traits.hpp>
 
 #include <boost/fusion/algorithm/iteration/for_each.hpp>
 #include <boost/iterator/iterator_facade.hpp>
@@ -528,7 +529,7 @@ namespace hpx { namespace traits
         {
             return segment_iterator(
                 functional::lift_zipped_iterators<
-                        functional::segmented_iterator_segment, iterator
+                        util::functional::segmented_iterator_segment, iterator
                     >::call(iter));
         }
 
@@ -538,7 +539,7 @@ namespace hpx { namespace traits
         {
             return local_iterator(
                 functional::lift_zipped_iterators<
-                        functional::segmented_iterator_local, iterator
+                        util::functional::segmented_iterator_local, iterator
                     >::call(iter));
         }
 
@@ -548,7 +549,7 @@ namespace hpx { namespace traits
         {
             return local_iterator(
                 functional::lift_zipped_iterators<
-                        functional::segmented_iterator_begin, iterator
+                        util::functional::segmented_iterator_begin, iterator
                     >::call(iter));
         }
 
@@ -558,7 +559,7 @@ namespace hpx { namespace traits
         {
             return local_iterator(
                 functional::lift_zipped_iterators<
-                        functional::segmented_iterator_end, iterator
+                        util::functional::segmented_iterator_end, iterator
                     >::call(iter));
         }
 
@@ -568,7 +569,8 @@ namespace hpx { namespace traits
         {
             return local_raw_iterator(
                 functional::lift_zipped_iterators<
-                        functional::segmented_iterator_local_begin, iterator
+                        util::functional::segmented_iterator_local_begin,
+                        iterator
                     >::call(seg_iter));
         }
 
@@ -578,7 +580,8 @@ namespace hpx { namespace traits
         {
             return local_raw_iterator(
                 functional::lift_zipped_iterators<
-                        functional::segmented_iterator_local_end, iterator
+                        util::functional::segmented_iterator_local_end,
+                        iterator
                     >::call(seg_iter));
         }
